fix(protocol): Move broadcasting of connected event to ConfigClient
The server doesn't know the client's caps until a bit after the
connection, when the client itself sends its first message, which is a
ConfigClient one.

In order to inform other clients that depend on the caps to know what
to expect from the newly connected client, the broadcasting of the
event of a new client was moved to the handler of the ConfigClient
message.

For that reason, the ChatServer.broadcast() function had to be updated
to take a JID to be excluded, to prevent the server from sending a
connected event to the user that just joined the session.