ui(refactor): rework websocket comms & work towards webrtc call
There was indeed a refactor of all the types of messages exchanged
between the UI and the websocket server.  All the types had their
shapes changed with the introduction of the protocol library.

Some simplifications were made and indirections removed, but the new
message types provided the same feature set we had before.  It's just
working better now.  I think the call itself isn't working because of
probably something wrong with the setup of the webrtc layer, and not
in the communication channel.