refactor(ui): start breaking the dreadful store.js into piece
This change sets up a new structure to allow the components to use
high level hooks closer to where features are implemented.  Some of
these hooks will depend on a useContext hook that has a provider setup
on the top-level application definition (index.jsx).  And that leaves
the context directory with the context and reducer definitions.  The
authentication was ported to this and it already feels waaay less
convoluted.