refactor(proto/capture/ui): Call protocol cleanup
The messages `Call{Answer,Offer}` were coalesced into `SDP`.  The
`NewIceCandidate` message was renamed to `ICE`, and a new message type
was introduced: `CallRequest`.

The the SDP payload caries the `type` field already, and is enough to
allow us to be a teeny bit more straightforward with.

The `CallRequest` is the most interesting change.  This allows the
frontend to ask the backend to start the call.