server: first pass on adding auth
Broken this down into this first step that introduces the jsonwebtoken
dependency.  But also, moved some of the server pieces to their own
files.  And the authentication was hooked up with the new dependency,
and the server was extended to exchange the `AuthPeer` message that
responds with the API in the `auth` module.  Notice that it currently
doesn't really check the password against anything yet.