ui: change domain to localhost, update deps, cry over webpack
After a lot of time wasted, I found out that now webpack requires a
bunch of new configuration in order to do the same thing it was doing.
I used the opportunity to move the cert configuration into
`webpack.config.js`.

Also, the `./webpack` script was removed in favor of `webpack serve`,
which still relies on `webpack-dev-server`, but it is less verbose.